Improve populating custom drive folders
For personal drive, `permissionId` is used, where `permissionId` is a string of digits. For all other drives or folders, mix of upper and lower letters, digits and symbols are used.

In practice, custom drive folders are only needed for personal drives but not for team drives or folders. The changes could save one click when navigating into team drives.